inference-monitor/pyproject.toml
[build-system]
requires = ["hatchling>=1.21"]
build-backend = "hatchling.build"

[project]
name = "sentinel-inference-monitor"
version = "0.1.0"
description = "Silent Data Corruption detection sidecar for GPU inference servers"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
authors = [
    { name = "SENTINEL Team" },
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "numpy>=1.26",
    "scipy>=1.12",
    "grpcio>=1.60",
    "grpcio-tools>=1.60",
    "pydantic>=2.0",
    "structlog>=23.0",
    "prometheus-client>=0.20",
    "uvloop>=0.19; sys_platform != 'win32'",
    "datasketch>=1.6",
    "aiohttp>=3.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "mypy>=1.8",
    "ruff>=0.2",
]

[project.scripts]
sentinel-inference-monitor = "sentinel_inference.monitor:main"

[tool.hatch.build.targets.wheel]
packages = ["src/sentinel_inference"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
